App/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.3
google-generativeai==0.7.1
flask-wtf==1.1.1
gunicorn==21.2.0
markupsafe==2.1.3
markdown2==2.4.10